requirements.txt
accelerate==0.18.0
evaluate==0.4.0
hydra-core==1.3.2
jsonlines==3.1.0
numpy==1.21.5
omegaconf==2.3.0
openai==0.28.1
Requests==2.31.0
scikit_learn==1.0.2
texttable==1.6.4
tqdm==4.62.3
transformers==4.26.1
tensorboard==2.11.2